Update Gemmini code
- Delete obsolute pldi22 and pldi24 tests
- Update Gemmini schedules in tests/asplos25
- Update gemmini-rocc-tests with v0.7.1 API
