+++ apps/gemmini/src/exo/conv.py
@@ -55,101 +55,6 @@ def conv_on_cpu_stride_1(
                     output[b, orow, ocol, och] = tmp_res2
 
 
-def split_fission_dim(conv):
-    conv = old_split(conv, "ocol", 16, ["ocol_o", "ocol_i"], tail="cut_and_guard")
-    conv = old_split(conv, "och", 16, ["och_o", "och_i"], perfect=True)
-    conv = old_split(conv, "kch", 16, ["kch_o", "kch_i"], perfect=True)
-    conv = old_reorder(conv, "ocol_i och_o")
-    conv = old_lift_alloc(conv, "res : _", n_lifts=3)
-    conv = old_fission_after(conv, "res[_] = _", n_lifts=3)
-    conv = old_fission_after(conv, "for krow in _:_", n_lifts=3)
-    conv = old_reorder(conv, "och_i krow")
-    conv = old_reorder(conv, "och_i kcol")
-    conv = old_reorder(conv, "och_i kch_o")
-    conv = old_reorder(conv, "ocol_i krow")
-    conv = old_reorder(conv, "ocol_i kcol")
-    conv = old_reorder(conv, "ocol_i kch_o")
-    conv = old_reorder(conv, "och_o krow")
-    conv = simplify(conv)
-    conv = old_lift_alloc(conv, "i_s : _", n_lifts=6)
-    conv = old_lift_alloc(conv, "w_s : _", n_lifts=1)
-    conv = old_lift_alloc(conv, "w_s : _", n_lifts=1, mode="col")
-    conv = old_reorder(conv, "och_o kcol")
-    conv = old_reorder(conv, "och_o kch_o")
-    conv = old_lift_alloc(conv, "w_s : _", n_lifts=3)
-    conv = old_fission_after(conv, "w_s = _", n_lifts=5)
-    conv = old_fission_after(conv, "i_s = _", n_lifts=5)
-
-    return conv
-
-
-def replace_div_part(conv):
-    conv = replace(conv, "for ocol_i in _:_ #0", ld_acc_i32_vector)
-    conv = old_reorder(conv, "och_i kch_i")
-    conv = old_reorder(conv, "och_o kch_i")
-    conv = replace(conv, "for kch_i in _:_ #0", ld_i8_block_id1)
-    conv = old_reorder(conv, "kch_o ocol_i")
-    conv = replace(conv, "for ocol_i in _:_ #0", ld_i8_block_id2)
-    conv = old_reorder(conv, "kch_i och_i")
-    conv = replace(conv, "for ocol_i in _:_ #0", matmul_acc_i8)
-    conv = replace(conv, "for ocol_i in _:_ #0", st_acc_i8)
-
-    return conv
-
-
-def replace_mod_part(conv):
-    conv = replace(conv, "for ocol_i in _:_ #0", ld_acc_i32_vector)
-    conv = old_reorder(conv, "och_i kch_i")
-    conv = replace(conv, "for kch_i in _:_ #0", ld_i8_block_id1)
-    conv = replace(conv, "for ocol_i in _:_ #0", ld_i8_block_id2)
-    conv = old_reorder(conv, "kch_i och_i")
-    conv = replace(conv, "for ocol_i in _:_ #0", matmul_acc_i8)
-    conv = replace(conv, "for ocol_i in _:_ #0", st_acc_i8)
-
-    return conv
-
-
-def inline_div_part(conv):
-    conv = inline_vector(conv)
-    conv = lift_config(conv, "config_ld_acc_i32_vector(_)")
-    conv = inline_ld_id1(conv)
-    conv = lift_config(conv, "config_ld_i8_id1(_)")
-    conv = inline_matmul(conv)
-    conv = lift_config(conv, "config_matmul(_)")
-    conv = inline_st(conv)
-    conv = lift_config(conv, "config_st_acc_i8(_)")
-
-    return conv
-
-
-def inline_mod_part(conv):
-    conv = inline_vector(conv)
-    conv = inline_ld_id1(conv)
-    conv = inline_matmul(conv)
-    conv = inline_st(conv)
-    conv = delete_config(conv, "config_ld_acc_i32_vector(_) #1")
-    conv = delete_config(conv, "config_ld_i8_id1(_) #1")
-    conv = delete_config(conv, "config_matmul(_) #1")
-    conv = delete_config(conv, "config_st_acc_i8(_) #1")
-    conv = simplify(conv)
-
-    return conv
-
-
-def set_gemm_memories(conv):
-    conv = set_memory(conv, "res", GEMM_ACCUM)
-    conv = set_memory(conv, "i_s", GEMM_SCRATCH)
-    conv = set_memory(conv, "w_s", GEMM_SCRATCH)
-    try:
-        conv = set_memory(conv, "res #1", GEMM_ACCUM)
-        conv = set_memory(conv, "i_s #1", GEMM_SCRATCH)
-        conv = set_memory(conv, "w_s #1", GEMM_SCRATCH)
-    except:
-        pass
-
-    return conv
-
-
 def schedule_conv_3():
     batch_size = 4
     out_channel = 64
