+++ apps/gemmini/test/bare-metal/shims/include/gemmini.h.in
@@ -5,30 +5,31 @@
 
 // Symbol overrides
 #define tiled_matmul_auto orig_tiled_matmul_auto
-#define tiled_conv_A_stride_auto orig_tiled_conv_A_stride_auto
+#define tiled_conv_auto orig_tiled_conv_auto
 
 #include "@gemmini-rocc-tests_SOURCE_DIR@/include/gemmini.h"
 
 #undef tiled_matmul_auto
-#undef tiled_conv_A_stride_auto
+#undef tiled_conv_auto
 
 void tiled_matmul_auto(size_t dim_I, size_t dim_J, size_t dim_K,
-    const elem_t *A, const elem_t *B, const void *D, void *C, size_t stride_A,
+    const elem_t* A, const elem_t* B, const void * D, void * C, size_t stride_A,
     size_t stride_B, size_t stride_D, size_t stride_C, scale_t A_scale_factor,
     scale_t B_scale_factor, scale_acc_t D_scale_factor, int act,
-    acc_scale_t scale, size_t relu6_shift, bool repeating_bias,
+    acc_scale_t scale, acc_scale_t bert_scale, bool repeating_bias,
     bool transpose_A, bool transpose_B, bool full_C, bool low_D,
     uint8_t weightA, enum tiled_matmul_type_t tiled_matmul_type);
 
-void tiled_conv_A_stride_auto(int batch_size, int in_dim,
-    int in_channels, int out_channels, int out_dim, int stride,
+void tiled_conv_auto( int batch_size, int in_row_dim, int in_col_dim,
+    int in_channels, int out_channels, int out_row_dim, int out_col_dim, int stride,
     int input_dilation, int kernel_dilation, int padding, int kernel_dim,
     bool wrot180, bool trans_output_1203, bool trans_input_3120,
     bool trans_weight_1203, bool trans_weight_0132,
-    const elem_t *input, const elem_t *weights, const acc_t *bias,
-    elem_t *output,
-    int act, acc_scale_t scale, size_t relu6_shift, int pool_size,
-    int pool_stride, int pool_padding,
+    const elem_t * input, const elem_t * weights, const acc_t * bias,
+    elem_t * output,
+    int act, acc_scale_t scale,
+    int pool_size, int pool_stride, int pool_padding,
     enum tiled_matmul_type_t tiled_conv_type);
+
 
 #endif
