+++ tests/golden/asplos25/test_higher_order/test_reduce.txt
