+++ tests/golden/test_apps/test_gemmini_matmul.txt
@@ -262,7 +262,7 @@ gemmini_extended3_config_ld({src_stride}, 1.0f, 0, 2);
 
 /* relying on the following instruction..."
 config_matmul()
-gemmini_extended_config_ex(WS, 0, 0, 0, 1, 0, 0);
+gemmini_extended_config_ex(WS, 0, 0, 1, 0, 0);
 
 */
 
@@ -535,7 +535,7 @@ gemmini_extended_mvin( 0, ((uint64_t) &{dst_data}),{m}, {n} );
 void matmul_14( test_case_Context *ctxt, const float* scale, bool act, const int8_t* A, const int8_t* B, int8_t* C ) {
 gemmini_extended_config_st((512), (act), (scale)[0]);
 
-gemmini_extended_config_ex(WS, 0, 0, 0, 1, 0, 0);
+gemmini_extended_config_ex(WS, 0, 0, 1, 0, 0);
 
 gemmini_extended3_config_ld((512), 1.0f, 0, 2);
 
@@ -632,7 +632,7 @@ gemm_acc_free((uint32_t)(res));
 void matmul_16( test_case_Context *ctxt, const float* scale, bool act, const int8_t* A, const int8_t* B, int8_t* C ) {
 gemmini_extended_config_st((128), (act), (scale)[0]);
 
-gemmini_extended_config_ex(WS, 0, 0, 0, 1, 0, 0);
+gemmini_extended_config_ex(WS, 0, 0, 1, 0, 0);
 
 gemmini_extended3_config_ld((128), 1.0f, 0, 2);
 
@@ -729,7 +729,7 @@ gemm_acc_free((uint32_t)(res));
 void matmul_27( test_case_Context *ctxt, const float* scale, bool act, const int8_t* A, const int8_t* B, int8_t* C ) {
 gemmini_extended_config_st((1024), (act), (scale)[0]);
 
-gemmini_extended_config_ex(WS, 0, 0, 0, 1, 0, 0);
+gemmini_extended_config_ex(WS, 0, 0, 1, 0, 0);
 
 gemmini_extended3_config_ld((1024), 1.0f, 0, 2);
 
@@ -822,7 +822,7 @@ gemm_acc_free((uint32_t)(res));
 void matmul_4( test_case_Context *ctxt, const float* scale, bool act, const int8_t* A, const int8_t* B, int8_t* C ) {
 gemmini_extended_config_st((256), (act), (scale)[0]);
 
-gemmini_extended_config_ex(WS, 0, 0, 0, 1, 0, 0);
+gemmini_extended_config_ex(WS, 0, 0, 1, 0, 0);
 
 gemmini_extended3_config_ld((256), 1.0f, 0, 2);
 
@@ -917,7 +917,7 @@ gemm_free((uint64_t)(a));
 void matmul_512x512x512( test_case_Context *ctxt, const float* scale, bool act, const int8_t* A, const int8_t* B, int8_t* C ) {
 gemmini_extended_config_st((512), (act), (scale)[0]);
 
-gemmini_extended_config_ex(WS, 0, 0, 0, 1, 0, 0);
+gemmini_extended_config_ex(WS, 0, 0, 1, 0, 0);
 
 gemmini_extended3_config_ld((512), 1.0f, 0, 2);
 
@@ -1020,7 +1020,7 @@ gemm_free((uint64_t)(a));
 void matmul_6( test_case_Context *ctxt, const float* scale, bool act, const int8_t* A, const int8_t* B, int8_t* C ) {
 gemmini_extended_config_st((64), (act), (scale)[0]);
 
-gemmini_extended_config_ex(WS, 0, 0, 0, 1, 0, 0);
+gemmini_extended_config_ex(WS, 0, 0, 1, 0, 0);
 
 gemmini_extended3_config_ld((64), 1.0f, 0, 2);
